Refactor bundling process, again.
Use a two pass bundling process, where the first pass is used for a better checksum of the code. The second pass is then only done if the manifest has changed and takes care of the transpiling.

Result should act similarly to how it did before, with the difference that environment changes and changes to other files in the bundle are now reflected in the manifest checksum.